refactor(terminal): bring three grandfathered terminal files back under their ceilings (#14959)
The ownership work pushed all three past ceilings a grandfathered file may not
cross. Nothing was deleted to hit a number -- each block moved to a sibling and
stays wired in, and every ceiling is re-lowered to the new size in both copies.

api/terminal.py 1504 -> 1300 (ceiling 1420 -> 1300): the SSH stub handler, its
session manager and the three websocket helpers move to api/terminal_ssh.py, the
same extraction terminal_handlers (#210) and terminal_tools (#185) already got.
The /ws/ssh/{host_id} route stays on the router and imports them back under this
module's own names, so api.terminal.ssh_terminal_manager still resolves for its
existing patch target.

api/agent_terminal.py 1115 -> 858 (ceiling 1112 -> 858): the five host-selection
endpoints and the pending-selection store nothing else touches become a
sub-router in api/agent_terminal_host_selection.py, included by the parent. All
18 registered paths are byte-identical before and after.

services/agent_terminal/service.py 1022 -> 969 (ceiling 1015 -> 969): the two
command-audit helpers read no service state beyond the logger handed to them, so
they move to the package's utils module as free functions and both call sites
follow.
