+++ scripts/python_file_size_known_large.py
@@ -51,7 +51,7 @@ KNOWN_LARGE: dict[str, int] = {
     "autobot-backend/api/advanced_control.py": 626,
     "autobot-backend/api/agent.py": 1380,
     "autobot-backend/api/agent_config.py": 1369,
-    "autobot-backend/api/agent_terminal.py": 1112,
+    "autobot-backend/api/agent_terminal.py": 858,
     "autobot-backend/api/ai_stack_integration.py": 638,
     "autobot-backend/api/analytics.py": 1288,
     "autobot-backend/api/analytics_architecture.py": 1264,
@@ -149,7 +149,7 @@ KNOWN_LARGE: dict[str, int] = {
     "autobot-backend/api/security_assessment.py": 913,
     "autobot-backend/api/settings.py": 1105,
     "autobot-backend/api/system.py": 1112,
-    "autobot-backend/api/terminal.py": 1420,
+    "autobot-backend/api/terminal.py": 1300,
     "autobot-backend/api/terminal_handlers.py": 1504,
     "autobot-backend/api/validation_dashboard.py": 671,
     "autobot-backend/api/vnc_manager.py": 1699,
@@ -328,7 +328,7 @@ KNOWN_LARGE: dict[str, int] = {
     "autobot-backend/security/threat_intelligence.py": 771,
     "autobot-backend/security_layer.py": 767,
     "autobot-backend/services/agent_analytics.py": 686,
-    "autobot-backend/services/agent_terminal/service.py": 1015,
+    "autobot-backend/services/agent_terminal/service.py": 969,
     "autobot-backend/services/ai_stack_client.py": 790,
     "autobot-backend/services/analytics_service.py": 838,
     "autobot-backend/services/audit_logger.py": 1022,
